requirements.txt
fastapi==0.110.0
sqlmodel==0.0.16
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1